garden_to_hpss: add fail safes

+++ emmet/scripts/garden_to_hpss.sh
@@ -7,9 +7,19 @@ for block_dir in `find $1 -maxdepth 1 -type d -name "block_*"`; do
   subdir=`basename $block_dir`
   if [ ! -e $1/archives/${subdir}.tar.gz ]; then
     tar -czvf $1/archives/${subdir}.tar.gz -C $1 $subdir
+    flag=$?
+    if [ $flag -ne 0 ]; then
+      echo "error with ${subdir}.tar.gz (flag=$flag)"
+      rm -v $1/archives/${subdir}.tar.gz
+      continue
+    fi
   fi
   hsi -l matcomp cput $1/archives/${subdir}.tar.gz : garden/${subdir}.tar.gz
   flag=$?
-  [[ $flag -ne 0 ]] && echo "not removing ${subdir}.tar.gz (flag=$flag)" && continue
+  if [ $flag -ne 0 ]; then
+    echo "error with hsi transfer for ${subdir}.tar.gz (flag=$flag)"
+    exit
+  fi
   rm -v $1/archives/${subdir}.tar.gz
+  rm -rv $block_dir
 done
